Refactoring BigInteger

+++ src/bench/scala/io/iohk/protocol/nizk/MultRelationNIZKPerformance.scala
@@ -1,7 +1,5 @@
 package io.iohk.protocol.nizk
 
-import java.math.BigInteger
-
 import io.iohk.core.crypto.encryption
 import io.iohk.core.crypto.encryption.Randomness
 import io.iohk.core.crypto.encryption.elgamal.{ElGamalCiphertext, LiftedElGamalEnc}
@@ -33,7 +31,7 @@ class MultRelationNIZKPerformance {
     for (size <- unitVectorSize) {
       println("Running test for unit vector of size " + size + " ...")
       val (uv, rand) = TimeUtils.time("\tUV creation: ", createUnitVector(size, 3))
-      val value = LiftedElGamalEnc.encrypt(pubKey, BigInteger.valueOf(5)).get._1
+      val value = LiftedElGamalEnc.encrypt(pubKey, 5).get._1
       val unitVector = for(i <- 0 until size) yield if(i == 3) BigInt(1) else BigInt(0)
       val uv2 = TimeUtils.time("\tUV with value creation: ",
         MultRelationNIZK.produceEncryptedUnitVectorWithValue(pubKey, value, unitVector))
@@ -52,7 +50,7 @@ class MultRelationNIZKPerformance {
     for (size <- unitVectorSize) {
       println("Running test for unit vector of size " + size + " ...")
       val (uv, rand) = TimeUtils.time("\tUV creation: ", createUnitVector(size, 3))
-      val value = LiftedElGamalEnc.encrypt(pubKey, BigInteger.valueOf(5)).get._1
+      val value = LiftedElGamalEnc.encrypt(pubKey, 5).get._1
       val unitVector = for(i <- 0 until size) yield if(i == 3) BigInt(1) else BigInt(0)
       val uv2 = TimeUtils.time("\tUV with value creation: ",
         MultRelationNIZK.produceEncryptedUnitVectorWithValue(pubKey, value, unitVector))
